fix: hang on n>1 libraries in the same directory
`LibraryObserver.stop()` mistakenly assumed that
`DirectoryObserver.stop()` would always stop the observer thread and
attempted to join() it, which caused an infinite hang when there were
multiple watched libraries in a common directory.

in such a case they are handled by the same observer thread, so calling
`DirectoryObserver.stop()` for just one of them (correctly) does not
stop the thread.

moved the join() into `DirectoryObserver.stop` where it belongs and
added a timeout for good measure. ideally now if anything goes wrong
with a watcher thread we don't bring the whole blender down.

fixes #5