fix: race condition on file reload
Reloading a file with linked libraries would cause a filesystem open
event on the library .blends, which would trigger observer callbacks at
the same time as observer stopping (and restarting) was taking place.

Specifically, since deleting the entry in the
`DirectoryObserver.watched_directories` dict happens before joining the
observer on line 122, we'd get a KeyError in call_callbacks:

https://github.com/gandalf3/auto-reload-linked-libraries/issues/6#issuecomment-2357210560

For now keep that behavior as is, since if filesystem events do happen
during observer stop we most likely don't want to run the callbacks
anyway, and the thread dying is inconsequential. If there are other
situations which cause this, I want to see the traceback.

Instead this fix just dodges the issue by ignoring file open events,
which should have been ignored anyway.

Perhaps it would be resonable avoid re-initializing observers on file
reload, but my intuition says to keep that behavior